Memperbarui File page admin dan utama bseserta logika hapus kamar

+++ hooks/useHapusKamar.js
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { doc, deleteDoc } from "firebase/firestore";
+import { doc, getDoc, deleteDoc } from "firebase/firestore";
 import { toast } from "react-toastify";
 // PERPUSTAKAAN KAMI
 import { database } from "@/lib/firebaseConfig";
@@ -10,7 +10,22 @@ const useHapusKamar = () => {
   const hapusKamar = async (id) => {
     try {
       setSedangMemuatHapusKamar(true);
+
+      // Ambil data kamar berdasarkan ID
       const referensiKamar = doc(database, "kamar", id);
+      const snapshot = await getDoc(referensiKamar);
+
+      if (!snapshot.exists()) {
+        throw new Error("Data kamar tidak ditemukan.");
+      }
+
+      const dataKamar = snapshot.data();
+
+      if (dataKamar.Status === "Terisi") {
+        toast.error("Kamar tidak dapat dihapus karena masih ada penghuninya");
+        return;
+      }
+
       await deleteDoc(referensiKamar);
       toast.success("Kamar berhasil dihapus!");
     } catch (error) {
